[GLES2] allow glShaderSource to deal with negative length items
The glShaderSource function should allow the length array passed to it
to contain negative values. In this case the length value should not be
used, instead the string with the same index should be considered as a
NULL-terminated string.

The game Taisei calls glShaderSource in this way (all length values are
-1). This commit makes this game running well with gl4es and Mesa GLES2
driver.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>