[GLES2] fix the place of #extension when GL_EXT_frag_depth is present
When GL_EXT_frag_depth is present, the shader converter will try to add
a #extension sentence to enable it. However, it's wrongly placed after a
line of ordinary definition in the header, which makes the converted shader
grammarly error.

Fix this by insert the #extension sentence before the definition.

The sentences of GL_EXT_frag_depth used to lack a newline, and it's also
added in this commit.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>